WIP: decoupling GameState and PlayerState
