--- Content/Blueprints/BP_PoolGameState.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:34010f9aa333d77661bdf4575d636b9ce2d00be7cda9f06315d48ec9b845e755
-size 14424
