--- Content/Blueprints/Player/BP_AimingComponent.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:c1a79c53e66d8a93201d3850ebfc48ff9d86f90d7bc9b0d6e931b365abd55da3
-size 318458
