--- Content/Blueprints/Player/BP_Billiardist.uasset
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:2e6d6a587f7326c68c8ff5561e4927a2e0603786cba20d5008dcf59faa2ab69d
-size 322682
