+++ Content/Blueprints/Player/Billiardist/BP_BilliardistPawn.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:54eecd0434e3956df2aefd541696bbcc10c50bc674b96982a0d4ccd97c1341a4
-size 150441
+oid sha256:519d1d10c57901846300068d0c50708754ff66afc99c1872b979f1ee9b0b9aa6
+size 146851
