+++ Source/Pool/AmericanPool/EightBallPlayerState.cpp
@@ -1,0 +1,5 @@
+// Copyright 2019 Andrei Vikarchuk.
+
+
+#include "EightBallPlayerState.h"
+
