+++ Source/Pool/AmericanPool/EightBallPlayerState.h
@@ -1,0 +1,17 @@
+// Copyright 2019 Andrei Vikarchuk.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "GameplayLogic/PoolPlayerState.h"
+#include "EightBallPlayerState.generated.h"
+
+/**
+ *
+ */
+UCLASS()
+class POOL_API AEightBallPlayerState : public APoolPlayerState
+{
+    GENERATED_BODY()
+
+};
