+++ Source/Pool/GameplayLogic/PoolGameMode.cpp
@@ -5,7 +5,7 @@
 #include "Objects/Table/Table.h"
 #include "Billiardist.h"
 #include "BilliardistController.h"
-#include "PoolGameState.h"
+#include "GameplayLogic/PoolGameState.h"
 
 #include "EngineUtils.h"
 #include "Kismet/KismetMathLibrary.h"
