+++ Source/Pool/GameplayLogic/RulesHandler.cpp
@@ -1,0 +1,8 @@
+// Copyright 2019 Andrei Vikarchuk.
+
+
+#include "RulesHandler.h"
+
+// Add default functionality here for any IRulesHandler functions that are not pure virtual.
+
+
